Move test and format security,js
